% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI.R
\docType{class}
\name{AOI}
\alias{AOI}
\title{Area of Interest class}
\description{
This class represents an area of interest for analysis.
}
\details{
See \code{\link[=aoi]{aoi()}} for details.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aux}}{The \link{CFAuxiliaryLongLat} instance using this AOI.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lonMin}}{Set or retrieve the lower longitude value.}

\item{\code{lonMax}}{Set or retrieve the higher longitude value.}

\item{\code{latMin}}{Set or retrieve the lower latitude value.}

\item{\code{latMax}}{Set or retrieve the higher latitude value.}

\item{\code{extent}}{Set of retrieve the four extremes of the AOI, a numeric
vector in the order longitude minimum, maximum, latitude minimum, maximum.}

\item{\code{resolution}}{Set or retrieve the resolution.}

\item{\code{dim}}{(read-only) The dimensions of the grid of the AOI once generated.}

\item{\code{dimnames}}{(read-only) Retrieve the dimnames of the AOI, in numeric
form. These are the center points of the grid cells.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AOI-new}{\code{AOI$new()}}
\item \href{#method-AOI-print}{\code{AOI$print()}}
\item \href{#method-AOI-clone}{\code{AOI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOI-new"></a>}}
\if{latex}{\out{\hypertarget{method-AOI-new}{}}}
\subsection{Method \code{new()}}{
Creating an instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOI$new(lonMin, lonMax, latMin, latMax, resolution)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lonMin, lonMax, latMin, latMax}}{The minimum and maximum values of the
longitude and latitude of the AOI, in decimal degrees. The longitude values
must agree with the range of the longitude in the variable to which this
AOI will be applied, e.g. \verb{[-180,180]} or \verb{[0,360]}.}

\item{\code{resolution}}{The separation between adjacent grid cell, in longitude
and latitude directions, in decimal degrees.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOI-print"></a>}}
\if{latex}{\out{\hypertarget{method-AOI-print}{}}}
\subsection{Method \code{print()}}{
Summary of the area of interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOI$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AOI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AOI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AOI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
