% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R
\name{axis,ncdfDimension-method}
\alias{axis,ncdfDimension-method}
\title{Dimension axis}
\usage{
\S4method{axis}{ncdfDimension}(x)
}
\arguments{
\item{x}{The \code{ncdfDimension} to get the axis for.}
}
\value{
One of \verb{X, Y, Z, T} or \code{NA_character_} when not known.
}
\description{
Dimension axis
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
time <- ds[["time"]]
axis(time)
}
