% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisScalar.R
\docType{class}
\name{CFAxisScalar}
\alias{CFAxisScalar}
\title{Scalar CF axis object}
\format{
An \link{R6Class} generator object.
}
\description{
This class represents a scalar axis. Its single value can be of
any type. It is typically used as an auxiliary axis to record some parameter
of interest such as the single time associated with a spatial grid with
longitude, latitude and vertical axes.
}
