% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFData.R
\docType{class}
\name{CFData}
\alias{CFData}
\title{Data extracted from a CF data variable}
\format{
An \link{R6Class} generator object.
}
\description{
This class holds the data that is extracted from a \link{CFVariable},
using the \code{subset()} class method. The instance of this class will
additionally have the axes and other relevant information such as the unit of
the data and the coordinate reference system.

The class has a number of utility functions to extract the data in specific
formats:
\itemize{
\item \code{raw()}: The array of the data without any further processing. The axes are
as they are stored in the netCDF resource; there is thus no guarantee as to
how the data is organized in the array. Dimnames will be set.
\item \code{array()}: An array of the data which is organized as a standard R array.
Dimnames will be set.
\item \code{stars()}: The data is returned as a \code{stars} object, with all relevant
metadata set. Package \code{stars} must be installed for this to work.
\item \code{rast()}: The data is returned as a \code{SpatRaster} object, with all relevant
metadata set. Package \code{terra} must be installed for this to work.
\item \code{data.table}: The data is returned as a \code{data.table}, with the first
columns corresponding to the dimension values of the axes of the data,
followed by a column "value". Package \code{data.table} must be installed for this
to work.
}
}
