% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{create_ncdf}
\alias{create_ncdf}
\title{Create a new data set}
\usage{
create_ncdf(fn)
}
\arguments{
\item{fn}{Optional. The fully qualified file name of the netCDF file if that
should be created. The file cannot already exist. It is recommended that
the filename uses an extension of ".nc". If the argument is not provided, a
virtual data set will be created.}
}
\description{
This function creates a new, empty data set with an associated netCDF file in
"netcdf4" format.
}
