% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOImethod.R
\name{dim.AOI}
\alias{dim.AOI}
\title{The dimensions of the grid of an AOI}
\usage{
\method{dim}{AOI}(x)
}
\arguments{
\item{x}{An instance of the \code{AOI} class.}
}
\value{
A vector of two values giving the longitude and latitude dimensions
of the grid that would be created for the AOI.
}
\description{
This method returns the dimensions of the grid that would be created for the
AOI.
}
\examples{
a <- aoi(30, 40, 10, 30, 0.1)
dim(a)
}
