% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFread.R
\name{.makeAxis}
\alias{.makeAxis}
\title{Create an \code{CFAxis} from an NC variable and dimension}
\usage{
.makeAxis(grp, var, dim)
}
\arguments{
\item{grp}{Group in which the NC variable is defined.}

\item{var}{\code{NCVariable} instance to create the axis from.}

\item{dim}{\code{NCDimension} associated with argument \code{var}.}
}
\value{
An instance of \code{CFAxis}.
}
\description{
Create an \code{CFAxis} from an NC variable and dimension
}
