% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFDataset.R
\docType{methods}
\name{groups}
\alias{groups}
\alias{groups.CFDataset}
\title{List the groups in the CF object, recursively.}
\usage{
groups(x)

\method{groups}{CFDataset}(x)
}
\arguments{
\item{x}{A \code{CFDataset} instance.}
}
\value{
A character vector with group names in the object.
}
\description{
List the groups in the CF object, recursively.
}
\examples{
fn <- system.file("extdata", "ERA5land_Rwanda_20160101.nc", package = "ncdfCF")
ds <- open_ncdf(fn)
groups(ds)
}
