% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R, R/ncdfDimensionNumeric.R,
%   R/ncdfDimensionTime.R
\name{has_bounds}
\alias{has_bounds}
\alias{has_bounds,ncdfDimension-method}
\alias{has_bounds,ncdfDimensionNumeric-method}
\alias{has_bounds,ncdfDimensionTime-method}
\title{Does the dimension have bounds?}
\usage{
has_bounds(x)

\S4method{has_bounds}{ncdfDimension}(x)

\S4method{has_bounds}{ncdfDimensionNumeric}(x)

\S4method{has_bounds}{ncdfDimensionTime}(x)
}
\arguments{
\item{x}{An instance of an \code{ncdfDimension} descendant class.}
}
\value{
\code{TRUE} if bounds are set, \code{FALSE} otherwise
}
\description{
This method indicates if the dimension has bounds set.
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
time <- ds[["time"]]
has_bounds(time)
}
