% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionTime.R
\name{indexOf,ANY,ncdfDimensionTime-method}
\alias{indexOf,ANY,ncdfDimensionTime-method}
\title{Find dimension indices in the time domain}
\usage{
\S4method{indexOf}{ANY,ncdfDimensionTime}(x, y, method = "constant")
}
\arguments{
\item{x}{Vector of character timestamps, \code{POSIXct} or \code{Date} values to find
indices for.}

\item{y}{An \code{ncdfDimensionTime} instance.}

\item{method}{Single value of "constant" or "linear". If \code{"constant"} or when
bounds are set on argument \code{y}, return the index value for each match. If
\code{"linear"}, return the index value with any fractional value.}
}
\value{
Numeric vector of the same length as \code{x}, or \code{NULL} if the range of
values in \code{x} is fully outside the valid range of \code{y}.
}
\description{
Find dimension indices in the time domain
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
time <- ds[["time"]]
indexOf(c("2024-03-01", "2024-03-02"), time)
}
