% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionNumeric.R
\name{indexOf,numeric,ncdfDimensionNumeric-method}
\alias{indexOf,numeric,ncdfDimensionNumeric-method}
\title{Find indices in the dimension domain}
\usage{
\S4method{indexOf}{numeric,ncdfDimensionNumeric}(x, y, method = "constant")
}
\arguments{
\item{x}{Vector of numeric values to find dimension indices for.}

\item{y}{An \code{ncdfDimensionNumeric} instance.}

\item{method}{Single value of "constant" or "linear". If \code{"constant"}, return
the index value for each match. If \code{"linear"}, return the index value with
any fractional value.}
}
\value{
Numeric vector of the same length as \code{x}.
}
\description{
Find indices in the dimension domain
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- ncdfDataset(fn)
lon <- ds[["lon"]]
indexOf(42:45, lon)
indexOf(42:45, lon, "linear")
}
