% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R
\name{length,ncdfDimension-method}
\alias{length,ncdfDimension-method}
\title{Length of the dimension}
\usage{
\S4method{length}{ncdfDimension}(x)
}
\arguments{
\item{x}{The \code{ncdfDimension} to query.}
}
\value{
Integer scalar of the length of the dimension.
}
\description{
Length of the dimension
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
time <- ds[["time"]]
length(time)
}
