% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeCharacterAxis}
\alias{makeCharacterAxis}
\title{Create a character axis}
\usage{
makeCharacterAxis(name, group, values, attributes = NULL)
}
\arguments{
\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{values}{The character coordinate values of the axis.}

\item{attributes}{\code{data.frame} with the attributes of the axis to create.}
}
\value{
A \link{CFAxisCharacter} instance.
}
\description{
With this method you can create a character axis to use with new \link{CFArray}
instances.
}
