% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeGroup}
\alias{makeGroup}
\title{Create a new group}
\usage{
makeGroup(name = "", parent_group = NULL)
}
\arguments{
\item{name}{A name for the group. This must be a valid name for CF objects.
The default value is the empty string "".}

\item{parent_group}{Optionally, a parent for the current group. This must be
an instance of \code{CFGroup} itself.}
}
\value{
A new instance of \code{CFGroup}.
}
\description{
This function creates a new group for CF objects. Groups are organized in an
hierarchy, starting with a root node. A root node is specified by an empty
string for a name and with argument \code{parent_group = NULL}.
}
\examples{
root <- makeGroup()
sub_group <- makeGroup("sub", root)
}
