% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeLongitudeAxis}
\alias{makeLongitudeAxis}
\title{Create a longitude axis}
\usage{
makeLongitudeAxis(
  name,
  values,
  bounds = NULL,
  attributes = data.frame(),
  group = NULL
)
}
\arguments{
\item{name}{Name of the axis.}

\item{values}{The coordinate values.}

\item{bounds}{The bounds of the coordinate values, or \code{NULL} if not
available.}

\item{attributes}{\code{data.frame} with the attributes of the axis to create.
Attributes "standard_name", "units", "actual_range" and "axis" will be set
or updated.}

\item{group}{\link{CFGroup} instance where the axis will be located. If \code{NULL}
(default), a private group will be created for the axis.}
}
\value{
A \link{CFAxisLongitude} instance.
}
\description{
With this method you can create a longitude axis to use with new \link{CFVariable}
instances.
}
