% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeMemoryGroup}
\alias{makeMemoryGroup}
\title{Create a group in memory to hold CF objects}
\usage{
makeMemoryGroup(id, name, fullname, title, history)
}
\arguments{
\item{id}{The id of the group.}

\item{name}{The name of the group.}

\item{fullname}{The full path and name of the group.}

\item{title}{The title attribute of the group.}

\item{history}{The history attribute of the group.}
}
\value{
A \code{MemoryGroup} instance.
}
\description{
Create a group in memory to hold CF objects
}
