% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeTimeAxis}
\alias{makeTimeAxis}
\title{Create a time axis}
\usage{
makeTimeAxis(name, group, values, attributes = NULL)
}
\arguments{
\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{values}{A \code{CFTime} instance with time values and optionally bounds set.}

\item{attributes}{\code{data.frame} with the attributes of the axis to create.
Attributes "standard_name", "units", "calendar", actual_range" and "axis"
will be set or updated.}
}
\value{
A \link{CFAxisTime} instance.
}
\description{
With this method you can create a time axis to use with new \link{CFArray}
instances.
}
