% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeVerticalAxis}
\alias{makeVerticalAxis}
\title{Create a vertical axis}
\usage{
makeVerticalAxis(name, values, bounds, attributes = NULL)
}
\arguments{
\item{name}{Name of the axis.}

\item{values}{The coordinate values.}

\item{bounds}{The bounds of the coordinate values, or \code{NULL} if not
available.}

\item{attributes}{\code{data.frame} with the attributes of the axis to create.
Attributes "actual_range" and "axis" will be set or updated.}
}
\value{
A \link{CFAxisVertical} instance.
}
\description{
With this method you can create a vertical axis to use with new \link{CFVariable}
instances. Note that you should set the "positive" attribute after creating
the axis to indicate if values are increasing going upwards (positive = "up")
or downwards (positive = "down").
}
