% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeVirtualGroup}
\alias{makeVirtualGroup}
\title{Create a group in memory to hold CF objects}
\usage{
makeVirtualGroup(id, name, fullname)
}
\arguments{
\item{id}{The id of the group.}

\item{name}{The name of the group.}

\item{fullname}{The full path and name of the group.}
}
\value{
A \code{VirtualGroup} instance.
}
\description{
With this function a group is created in memory, i.e. not associated with a
netCDF resource on file. This can be used to prepare new CF objects before
writing them to file. Extracting data from a \code{CFVariable} into a \link{CFArray}
instance will also create a virtual group.
}
