% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfObject.R
\name{name,ncdfObject-method}
\alias{name,ncdfObject-method}
\title{Retrieve the name of an ncdfCF object}
\usage{
\S4method{name}{ncdfObject}(object)
}
\arguments{
\item{object}{The object whose name to retrieve.}
}
\value{
A character string of the name of the object.
}
\description{
Retrieve the name of an ncdfCF object
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
name(ds[["lon"]])
}
