% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDataset.R
\name{names,ncdfDataset-method}
\alias{names,ncdfDataset-method}
\title{Variable names of an \code{ncdfDataset} instance}
\usage{
\S4method{names}{ncdfDataset}(x)
}
\arguments{
\item{x}{\code{ncdfDataset} whose variable names to retrieve.}
}
\value{
A character vector of variable names.
}
\description{
Variable names of an \code{ncdfDataset} instance
}
\examples{
fn <- system.file("extdata",
  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
  package = "ncdfCF")
ds <- open_ncdf(fn)
names(ds)
}
