% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDataset.R
\docType{class}
\name{ncdfDataset-class}
\alias{ncdfDataset-class}
\title{ncdfDataset class}
\description{
This class represents a single NetCDF resource.
}
\section{Slots}{

\describe{
\item{\code{resource}}{The \code{ncdfResource} instance that handles the NetCDF file.}

\item{\code{keep_open}}{Logical flag to indicate if the resource should remain open
for data access after the initial read of metadata.}

\item{\code{vars}}{A list with the variables in the resource.}

\item{\code{dims}}{A list holding all the dimension data.}

\item{\code{format}}{A character string with the format of the NetCDF resource.}

\item{\code{has_error}}{logical. Flag to indicate if there was an error opening the resource.}
}}

