% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R
\docType{class}
\name{ncdfDimension-class}
\alias{ncdfDimension-class}
\title{Base dimension object}
\description{
The 'dimension' is one of the key building blocks of a data set in an netCDF
resource. This virtual class contains the common functionality of all
dimension classes. Instantiable sub-classes include \code{ncdfDimensionNumeric} for
numeric dimension values, and \code{ncdfDimensionTime} for "time"
dimensions.
}
\section{Slots}{

\describe{
\item{\code{var_id}}{The ID of the dimension variable if read from file.}

\item{\code{var_type}}{The type of the dimension variable if read from file.}

\item{\code{length}}{The number of elements in the dimension.}

\item{\code{unlim}}{Flag to indicate whether this dimension is unlimited, e.g. can be
extended beyond the current 'length'.}

\item{\code{axis}}{The axis of the dimension, if known.}
}}

