% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionCharacter.R
\docType{class}
\name{ncdfDimensionCharacter-class}
\alias{ncdfDimensionCharacter-class}
\title{Character dimension class}
\description{
This class describes character dimensions in an NetCDF resource. Character
dimensions are discrete by default.
}
\section{Slots}{

\describe{
\item{\code{values}}{The values of the positions along the dimension.}
}}

