% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionDiscrete.R
\docType{class}
\name{ncdfDimensionDiscrete-class}
\alias{ncdfDimensionDiscrete-class}
\title{Discrete dimension class}
\description{
This class describes discrete dimensions in an NetCDF resource. A discrete
dimension has no values. The indices along the dimension run from 1 to the
length of the dimension.
}
