% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R
\name{ncdfDimensionGenerics}
\alias{ncdfDimensionGenerics}
\alias{has_bounds}
\alias{axis}
\title{Generics for \code{ncdfCF} dimensions}
\usage{
has_bounds(x)

axis(x)
}
\arguments{
\item{x}{The \code{ncdfCF} dimension that the method operates on.}
}
\description{
These are generic method definitions with implementations in descendant
classes. See \code{ncdfDimensionNumeric} and \code{ncdfDimensionTime} help topics for
details.
}
