% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionTime.R
\docType{class}
\name{ncdfDimensionTime-class}
\alias{ncdfDimensionTime-class}
\title{Dimension object}
\description{
The 'dimension' is one of the key building blocks of a data set in an netCDF
resource. This class is specifically for dimensions that represent the "time"
dimension.
}
\section{Slots}{

\describe{
\item{\code{values}}{The values of the positions along the dimension, as an instance
of S4 class 'CFtime'.}
}}

