% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfObject.R
\name{ncdfGenerics}
\alias{ncdfGenerics}
\alias{id}
\alias{name}
\alias{attribute}
\alias{show_attributes}
\title{Generics for \code{ncdfCF} objects}
\usage{
id(object)

name(object)

attribute(object, att, field = "value")

show_attributes(object, ...)
}
\arguments{
\item{object}{The \code{ncdfCF} object that the method operates on. This includes
datasets, variables, dimensions, and possible others including instances
of descendant classes.}

\item{att}{Name of the attribute to look up.}

\item{field}{One of "type", "length" or "value" (default)}

\item{...}{Passed on to underlying functions.}
}
\value{
Various. Please see the documentation of the methods in descendant
classes.
}
\description{
These are generic method definitions with implementations in descendant
classes. Please see the help topics for the descendant classes.
}
