% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfGroup.R
\docType{class}
\name{ncdfGroup-class}
\alias{ncdfGroup-class}
\title{Group class}
\description{
This class represents a group in a NetCDF resource. This applies equally to
the "classic" data model, which will have only a root group. Resources in the
"netcdf4" format can have multiple, hierarchical groups.
}
\section{Slots}{

\describe{
\item{\code{resource}}{The \code{ncdfResource} instance that handles the NetCDF file.}

\item{\code{handle}}{The RNetCDF handle to the group.}

\item{\code{parent}}{The parent group of this group.}

\item{\code{subgroups}}{The subgroups of this group, if any.}

\item{\code{vars}}{A list with the variables in the resource.}

\item{\code{dims}}{A list holding all the dimension data.}
}}

