% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfObject.R
\docType{class}
\name{ncdfObject-class}
\alias{ncdfObject-class}
\title{Ancestor of all netCDF objects.}
\description{
This is a virtual class that is the ancestor of other ncdfCF classes. Slots
'id' and 'name' are populated in the descending classes.
}
\section{Slots}{

\describe{
\item{\code{id}}{Identifier of the netCDF object.}

\item{\code{name}}{Name of the netCDF object.}

\item{\code{attributes}}{A data.frame holding all attributes of the object.}

\item{\code{resource}}{\code{ncdfResource} instance to access the netCDF resource.}
}}

