% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfResource.R
\name{ncdfOpenClose}
\alias{ncdfOpenClose}
\alias{open,ncdfResource-method}
\alias{close,ncdfResource-method}
\title{Open or close a NetCDF resource}
\usage{
\S4method{open}{ncdfResource}(con, ...)

\S4method{close}{ncdfResource}(con, ...)
}
\arguments{
\item{con}{The \code{ncdfResource} instance of the \code{ncdfDataset} to be opened or
closed.}

\item{...}{Ignored.}
}
\value{
The \code{open()} method returns a handle to the NetCDF resource, or an
error. Method \code{close()} returns nothing.
}
\description{
These methods open or close a NetCDF resource. It is typically not necessary
to call these functions directly, the \code{ncdfCF} objects call these functions
at appropriate times.
}
