% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfResource.R
\docType{class}
\name{ncdfResource-class}
\alias{ncdfResource-class}
\title{Low-level access to the RNetCDF package}
\description{
This class, of which there will be 1 instance in the package, shared by all
\code{ncdfObject}s, provides low-level access to the \code{RNetCDF} package, mainly to
provide a valid handle with which to access the underlying netCDF resource.
}
\section{Slots}{

\describe{
\item{\code{uri}}{character.}

\item{\code{handle}}{ANY.}
}}

