% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfResource.R
\name{ncdfResource}
\alias{ncdfResource}
\title{Create an ncdfResource instance}
\usage{
ncdfResource(uri)
}
\arguments{
\item{uri}{The URI of the resource}
}
\value{
An \code{ncdfResource} instance with the underlying resource opened, or an
error message.
}
\description{
This function will create an \code{ncdfResource} instance with the underlying netCDF
resource opened.
}
