% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDataset.R
\name{objects_by_standard_name}
\alias{objects_by_standard_name}
\alias{objects_by_standard_name,ncdfDataset-method}
\title{Get objects by standard_name}
\usage{
objects_by_standard_name(x, standard_name)

\S4method{objects_by_standard_name}{ncdfDataset}(x, standard_name)
}
\arguments{
\item{x}{The \code{ncdfDataset} to query for the "standard_name" attribute.}

\item{standard_name}{A character string to search for in variables and
dimensions.}
}
\value{
A character vector of variable or dimension names.
}
\description{
Several conventions define standard vocabularies for physical properties. The
standard names from those vocabularies are usually stored as the
"standard_name" attribute with variables or dimensions. This methods
retrieves all variables or dimensions that list the specified "standard_name"
in its attributes.
}
\examples{
fn <- system.file("extdata",
  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
  package = "ncdfCF")
ds <- open_ncdf(fn)
objects_by_standard_name(ds, "precipitation_flux")
}
