% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfObject.R
\name{show_attributes,ncdfObject-method}
\alias{show_attributes,ncdfObject-method}
\title{Print the attributes of the object to the console}
\usage{
\S4method{show_attributes}{ncdfObject}(object)
}
\arguments{
\item{object}{\code{ncdfObject} whose attributes to print.}
}
\value{
Nothing.
}
\description{
Print the attributes of the object to the console
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
show_attributes(ds[["pr"]])
}
