% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxis.R
\name{str.CFAxis}
\alias{str.CFAxis}
\title{Compact display of an axis.}
\usage{
\method{str}{CFAxis}(object, ...)
}
\arguments{
\item{object}{A \code{CFAxis} instance or any descendant.}

\item{...}{Ignored.}
}
\description{
Compact display of an axis.
}
