% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFDataset.R
\name{str.CFDataset}
\alias{str.CFDataset}
\title{Compact display of a CFDataset}
\usage{
\method{str}{CFDataset}(object, ...)
}
\arguments{
\item{object}{A \code{CFDataset} instance.}

\item{...}{Ignored.}
}
\description{
Compact display of a CFDataset
}
