% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfObject.R, R/ncdfVariable.R, R/ncdfGroup.R,
%   R/ncdfDataset.R, R/ncdfDimensionCharacter.R, R/ncdfDimensionDiscrete.R,
%   R/ncdfDimensionNumeric.R, R/ncdfDimensionTime.R
\name{str}
\alias{str}
\alias{str,ncdfVariable-method}
\alias{str,ncdfGroup-method}
\alias{str,ncdfDataset-method}
\alias{str,ncdfDimensionCharacter-method}
\alias{str,ncdfDimensionDiscrete-method}
\alias{str,ncdfDimensionNumeric-method}
\alias{str,ncdfDimensionTime-method}
\title{\code{str} for \code{ncdfCF} objects}
\usage{
str(object, ...)

\S4method{str}{ncdfVariable}(object, ...)

\S4method{str}{ncdfGroup}(object, ...)

\S4method{str}{ncdfDataset}(object, ...)

\S4method{str}{ncdfDimensionCharacter}(object, ...)

\S4method{str}{ncdfDimensionDiscrete}(object, ...)

\S4method{str}{ncdfDimensionNumeric}(object, ...)

\S4method{str}{ncdfDimensionTime}(object, ...)
}
\arguments{
\item{object}{The \code{ncdfObject} that the method operates on. This includes
datasets, variables, dimensions, and possible others including instances
of descendant classes.}

\item{...}{Ignored.}
}
\value{
Character vector that describes the structure of the object.
}
\description{
\code{ncdfObject}s have references back to their parent objects, specifically
\code{ncdfDataset} and \code{ncdfGroup}. These circular references lead to very bulky
displays of the structure of the object being examined when using the base R
version of \code{str()}. This package-specific version puts a short-circuit for
parent references, making the display more manageable, including in the
Environment pane of RStudio when in an interactive session.
}
\examples{
fn <- system.file("extdata", "ERA5land_Rwanda_20160101.nc", package = "ncdfCF")
ds <- open_ncdf(fn)
str(ds)
}
