% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimension.R
\name{time,ncdfDimension-method}
\alias{time,ncdfDimension-method}
\title{Get the full time specification of the dimension}
\usage{
\S4method{time}{ncdfDimension}(x)
}
\arguments{
\item{x}{A \code{ncdfDimension} instance.}
}
\value{
\code{NULL}
}
\description{
This method returns \code{NULL}. Class \code{ncdfDimensionTime} implements this method
more usefully.
}
\examples{
fn <- system.file("extdata",
                  "pr_day_EC-Earth3-CC_ssp245_r1i1p1f1_gr_20240101-20241231_vncdfCF.nc",
                  package = "ncdfCF")
ds <- open_ncdf(fn)
time(ds[["lon"]])
}
