% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAuxiliaryLongLat.R
\docType{class}
\name{CFAuxiliaryLongLat}
\alias{CFAuxiliaryLongLat}
\title{CF auxiliary longitude-latitude variable}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
This class represents the longitude and latitude variables that
compose auxiliary coordinate variable axes for X-Y grids that are not
longitude-latitude.
}
\details{
The class provides access to the data arrays for longitude and
latitude from the netCDF resource, as well as all the details that have been
associated with both axes. Additionally, this class can generate the index
to extract values on a long-lat grid of the associated X-Y grid data variable
using a user-selectable extent and resolution.
}
