% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisLatitude.R
\docType{class}
\name{CFAxisLatitude}
\alias{CFAxisLatitude}
\title{Latitude CF axis object}
\format{
An \link{R6Class} generator object.
}
\description{
This class represents a latitude axis. Its values are numeric.
This class is used for axes that represent latitudes.

This class adds some logic that is specific to latitudes, such as their
range, orientation and their meaning.
}
