% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisLongitude.R
\docType{class}
\name{CFAxisLongitude}
\alias{CFAxisLongitude}
\title{Longitude CF axis object}
\description{
This class represents a longitude axis. Its values are numeric.
This class is used for axes that represent longitudes. This class adds some
logic that is specific to longitudes, such as their range, orientation and
their meaning. (In the near future, it will also support selecting data
that crosses the 0-360 degree boundary.)
}
\section{Super classes}{
\code{\link[ncdfCF:CFObject]{CFObject}} -> \code{\link[ncdfCF:CFData]{CFData}} -> \code{\link[ncdfCF:CFAxis]{CFAxis}} -> \code{\link[ncdfCF:CFAxisNumeric]{CFAxisNumeric}} -> \code{CFAxisLongitude}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{friendlyClassName}}{(read-only) A nice description of the class.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CFAxisLongitude-initialize}{\code{CFAxisLongitude$new()}}
    \item \href{#method-CFAxisLongitude-copy}{\code{CFAxisLongitude$copy()}}
    \item \href{#method-CFAxisLongitude-copy_with_values}{\code{CFAxisLongitude$copy_with_values()}}
    \item \href{#method-CFAxisLongitude-subset}{\code{CFAxisLongitude$subset()}}
    \item \href{#method-CFAxisLongitude-append}{\code{CFAxisLongitude$append()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="append_attribute"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-append_attribute'><code>CFObject$append_attribute()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="attribute"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-attribute'><code>CFObject$attribute()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="attributes_identical"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-attributes_identical'><code>CFObject$attributes_identical()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="delete_attribute"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-delete_attribute'><code>CFObject$delete_attribute()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="print_attributes"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-print_attributes'><code>CFObject$print_attributes()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="set_attribute"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-set_attribute'><code>CFObject$set_attribute()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="write_attributes"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-write_attributes'><code>CFObject$write_attributes()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFData" data-id="dim"><a href='../../ncdfCF/html/CFData.html#method-CFData-dim'><code>CFData$dim()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFData" data-id="read_chunk"><a href='../../ncdfCF/html/CFData.html#method-CFData-read_chunk'><code>CFData$read_chunk()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFData" data-id="read_data"><a href='../../ncdfCF/html/CFData.html#method-CFData-read_data'><code>CFData$read_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="attach_to_group"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-attach_to_group'><code>CFAxis$attach_to_group()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="can_append"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-can_append'><code>CFAxis$can_append()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="configure_terms"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-configure_terms'><code>CFAxis$configure_terms()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="copy_terms"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-copy_terms'><code>CFAxis$copy_terms()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="detach"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-detach'><code>CFAxis$detach()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="peek"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-peek'><code>CFAxis$peek()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="shard"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-shard'><code>CFAxis$shard()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="write"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-write'><code>CFAxis$write()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="brief"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-brief'><code>CFAxisNumeric$brief()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="identical"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-identical'><code>CFAxisNumeric$identical()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="indexOf"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-indexOf'><code>CFAxisNumeric$indexOf()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="print"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-print'><code>CFAxisNumeric$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="range"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-range'><code>CFAxisNumeric$range()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxisNumeric" data-id="slice"><a href='../../ncdfCF/html/CFAxisNumeric.html#method-CFAxisNumeric-slice'><code>CFAxisNumeric$slice()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisLongitude-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisLongitude-initialize}{}}}
\subsection{\code{CFAxisLongitude$new()}}{
  Create a new instance of this class.

Creating a new longitude axis is more easily done with the
\code{\link[=makeLongitudeAxis]{makeLongitudeAxis()}} function.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CFAxisLongitude$new(
  var,
  group,
  values,
  start = 1L,
  count = NA,
  attributes = data.frame()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{var}}{The name of the axis when creating a new axis. When reading an
axis from file, the \link{NCVariable} object that describes this instance.}
      \item{\code{group}}{The \link{CFGroup} that this instance will live in.}
      \item{\code{values}}{Optional. The values of the axis in a vector. The values
have to be numeric with the maximum value no larger than the minimum
value + 360, and monotonic. Ignored when argument \code{var} is a NCVariable
object.}
      \item{\code{start}}{Optional. Integer index where to start reading axis data
from file. The index may be \code{NA} to start reading data from the start.}
      \item{\code{count}}{Optional. Number of elements to read from file. This may be
\code{NA} to read to the end of the data.}
      \item{\code{attributes}}{Optional. A \code{data.frame} with the attributes of the
axis. When an empty \code{data.frame} (default) and argument \code{var} is an
NCVariable instance, attributes of the axis will be taken from the
netCDF resource.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisLongitude-copy"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisLongitude-copy}{}}}
\subsection{\code{CFAxisLongitude$copy()}}{
  Create a copy of this axis. The copy is completely separate
from \code{self}, meaning that both \code{self} and all of its components are made
from new instances.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CFAxisLongitude$copy(name = "", group)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name for the new axis. If an empty string is passed, will
use the name of this axis.}
      \item{\code{group}}{The \link{CFGroup} where the copy of this axis will live.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The newly created axis.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisLongitude-copy_with_values"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisLongitude-copy_with_values}{}}}
\subsection{\code{CFAxisLongitude$copy_with_values()}}{
  Create a copy of this axis but using the supplied values.
The attributes are copied to the new axis. Boundary values and
auxiliary coordinates are not copied.

After this operation the attributes of the newly created axes may not
be accurate, except for the "actual_range" attribute. The calling code
should set, modify or delete attributes as appropriate.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CFAxisLongitude$copy_with_values(name = "", group, values)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name for the new axis. If an empty string is passed, will
use the name of this axis.}
      \item{\code{group}}{The \link{CFGroup} where the copy of this axis will live.}
      \item{\code{values}}{The values to the used with the copy of this axis.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The newly created axis.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisLongitude-subset"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisLongitude-subset}{}}}
\subsection{\code{CFAxisLongitude$subset()}}{
  Return a longitude axis spanning a smaller coordinate range.
This method returns an axis which spans the range of indices given by
the \code{rng} argument.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CFAxisLongitude$subset(name = "", group, rng = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name for the new axis. If an empty string is passed
(default), will use the name of this axis.}
      \item{\code{group}}{The \link{CFGroup} where the copy of this axis will live.}
      \item{\code{rng}}{The range of indices whose values from this axis to include in
the returned axis. If the value of the argument is \code{NULL}, return a
copy of the axis.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{CFAxisLongitude} instance covering the indicated range of
indices. If the value of the argument \code{rng} is \code{NULL}, return a copy of
\code{self} as the new axis.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisLongitude-append"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisLongitude-append}{}}}
\subsection{\code{CFAxisLongitude$append()}}{
  Append a vector of values at the end of the current values
of the axis. Boundary values are appended as well but if either this
axis or the \code{from} axis does not have boundary values, neither will the
resulting axis.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CFAxisLongitude$append(from)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{from}}{An instance of \code{CFAxisLongitude} whose values to append to
the values of this axis.}
      \item{\code{group}}{The \link{CFGroup} where the copy of this axis will live.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new \code{CFAxisLongitude} instance with values from this axis and
the \code{from} axis appended.
  }
}

}
