% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisLongitude.R
\docType{class}
\name{CFAxisLongitude}
\alias{CFAxisLongitude}
\title{Longitude CF axis object}
\format{
An \link{R6Class} generator object.
}
\description{
This class represents a longitude axis. Its values are numeric.
This class is used for axes that represent longitudes.

This class adds some logic that is specific to longitudes, such as their
range, orientation and their meaning. (In the near future, it will also
support selecting data that crosses the 0-360 degree boundary.)
}
