% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisNumeric.R
\docType{class}
\name{CFAxisNumeric}
\alias{CFAxisNumeric}
\title{Numeric CF axis object}
\format{
An \link{R6Class} generator object.
}
\description{
This class represents a numeric axis. Its values are numeric.
This class is used for axes with numeric values but without further
knowledge of their nature. More specific classes descend from this class.
}
