% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisTime.R
\docType{class}
\name{CFAxisTime}
\alias{CFAxisTime}
\title{Time axis object}
\description{
This class represents a time axis. The functionality is provided
by the \code{CFtime} package.
}
\section{Super classes}{
\code{\link[ncdfCF:CFObject]{ncdfCF::CFObject}} -> \code{\link[ncdfCF:CFAxis]{ncdfCF::CFAxis}} -> \code{CFAxisTime}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{The CFtime instance to manage CF time.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{friendlyClassName}}{(read-only) A nice description of the class.}

\item{\code{dimnames}}{(read-only) The coordinates of the axis as a character
vector.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CFAxisTime-new}{\code{CFAxisTime$new()}}
\item \href{#method-CFAxisTime-print}{\code{CFAxisTime$print()}}
\item \href{#method-CFAxisTime-brief}{\code{CFAxisTime$brief()}}
\item \href{#method-CFAxisTime-time}{\code{CFAxisTime$time()}}
\item \href{#method-CFAxisTime-indexOf}{\code{CFAxisTime$indexOf()}}
\item \href{#method-CFAxisTime-sub_axis}{\code{CFAxisTime$sub_axis()}}
\item \href{#method-CFAxisTime-clone}{\code{CFAxisTime$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="attribute"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-attribute'><code>ncdfCF::CFObject$attribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFObject" data-id="print_attributes"><a href='../../ncdfCF/html/CFObject.html#method-CFObject-print_attributes'><code>ncdfCF::CFObject$print_attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ncdfCF" data-topic="CFAxis" data-id="shard"><a href='../../ncdfCF/html/CFAxis.html#method-CFAxis-shard'><code>ncdfCF::CFAxis$shard()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$new(grp, nc_var, nc_dim, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grp}}{The group that contains the netCDF variable.}

\item{\code{nc_var}}{The netCDF variable that describes this instance.}

\item{\code{nc_dim}}{The netCDF dimension that describes the dimensionality.}

\item{\code{values}}{The CFtime instance that manages this axis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-print"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-print}{}}}
\subsection{Method \code{print()}}{
Summary of the time axis

Prints a summary of the time axis to the console.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-brief"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-brief}{}}}
\subsection{Method \code{brief()}}{
Retrieve a 1-row \code{data.frame} with some information on this axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$brief()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-time"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-time}{}}}
\subsection{Method \code{time()}}{
Retrieve the CFtime instance that manages this axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-indexOf"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-indexOf}{}}}
\subsection{Method \code{indexOf()}}{
Retrieve the indices of supplied values on the time axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$indexOf(x, method = "constant", rightmost.closed = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of timestamps whose indices into the time axis to
extract.}

\item{\code{method}}{Extract index values without ("constant", the default) or
with ("linear") fractional parts.}

\item{\code{rightmost.closed}}{Whether or not to include the upper limit.
Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer vector giving the indices in the time axis of valid
values in \code{x}, or \code{integer(0)} if none of the values are valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-sub_axis"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-sub_axis}{}}}
\subsection{Method \code{sub_axis()}}{
Return an axis spanning a smaller dimension range

This method returns an axis which spans the range of indices given by
the \code{rng} argument.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$sub_axis(group, rng = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{The group to create the new axis in.}

\item{\code{rng}}{The range of values from this axis to include in the returned
axis.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{CFAxisTime} covering the indicated range of indices. If the
\code{rng} argument includes only a single value, an \link{CFAxisScalar} instance
is returned with its value being the character timestamp of the value
in this axis. If the value of the argument is \code{NULL}, return the entire
axis (possibly as a scalar axis).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CFAxisTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CFAxisTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CFAxisTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
