% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFAxisVertical.R
\docType{class}
\name{CFAxisVertical}
\alias{CFAxisVertical}
\title{Parametric vertical CF axis object}
\format{
An \link{R6Class} generator object.
}
\description{
This class represents a parametric vertical axis. It is defined
through an index value that is contained in the axis, with additional
\link{NCVariable}s that hold ancillary data with which to calculate dimensional
axis values. It is used in atmosphere and ocean data sets. Non-parametric
vertical axes are stored in an \link{CFAxisNumeric}.
}
\references{
https://cfconventions.org/Data/cf-conventions/cf-conventions-1.11/cf-conventions.html#parametric-vertical-coordinate
}
