% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFDataset.R
\docType{class}
\name{CFDataset}
\alias{CFDataset}
\title{CF data set}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
This class represents a CF data set, the object that
encapsulates a netCDF resource. You should never have to instantiate this
class directly; instead, call \code{\link[=open_ncdf]{open_ncdf()}} which will return an instance
that has all properties read from the netCDF resource. Class methods can then
be called, or the base R functions called with this instance.
}
\details{
The CF data set instance provides access to all the objects in the
netCDF resource, organized in groups.
}
