% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFVariable.R
\docType{class}
\name{CFVariable}
\alias{CFVariable}
\title{CF data variable}
\format{
An \code{\link{R6Class}} generator object.
}
\description{
This class represents a CF data variable, the object that
provides access to an array of data.
}
\details{
The CF data variable instance provides access to the data array from
the netCDF resource, as well as all the details that have been associated
with the data variable, such as axis information, grid mapping parameters,
etc.
}
\section{Fields}{

\describe{
\item{\code{group}}{The \link{NCGroup} that this variable is defined in.}

\item{\code{axes}}{List of instances of classes descending from \link{CFAxis} that are
the axes of the variable.}
}}

