% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{create_ncdf}
\alias{create_ncdf}
\title{Create a new data set}
\usage{
create_ncdf()
}
\description{
This function creates a new, empty data set in memory. You can add groups,
data variables and attributes to this data set as appropriate. The data set
can be saved with \code{CFDataset$save()}, if so desired.
}
