% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeCharacterAxis}
\alias{makeCharacterAxis}
\title{Create a character axis}
\usage{
makeCharacterAxis(name, values, attributes = data.frame(), group = NULL)
}
\arguments{
\item{name}{Name of the axis.}

\item{values}{The character coordinate values of the axis.}

\item{attributes}{\code{data.frame} with the attributes of the axis to create.}

\item{group}{\link{CFGroup} instance where the axis will be located. If \code{NULL}
(default), a private group will be created for the axis.}
}
\value{
A \link{CFAxisCharacter} instance.
}
\description{
With this method you can create a character axis to use with new \link{CFVariable}
instances.
}
