% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeDiscreteAxis}
\alias{makeDiscreteAxis}
\title{Create a discrete axis}
\usage{
makeDiscreteAxis(name, length)
}
\arguments{
\item{name}{Name of the axis.}

\item{length}{The length of the axis.}
}
\value{
A \link{CFAxisDiscrete} instance. The values will be a sequence of size
\code{length}.
}
\description{
With this method you can create a discrete axis to use with new \link{CFVariable}
instances.
}
