% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeLatitudeAxis}
\alias{makeLatitudeAxis}
\title{Create a latitude axis}
\usage{
makeLatitudeAxis(id, name, group, values, bounds)
}
\arguments{
\item{id}{Id of the axis.}

\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{values}{The coordinate values.}

\item{bounds}{The bounds of the coordinate values, or \code{NULL} if not available.}
}
\value{
A \code{CFAxisLatitude} or \code{CFAxisScalar} instance.
}
\description{
With this method you can create a latitude axis to use with new \link{CFArray}
instances.
}
