% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeLongitudeAxis}
\alias{makeLongitudeAxis}
\title{Create a longitude axis}
\usage{
makeLongitudeAxis(name, group, values, bounds = NULL)
}
\arguments{
\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{values}{The coordinate values.}

\item{bounds}{The bounds of the coordinate values, or \code{NULL} if not
available.}
}
\value{
A \link{CFAxisLongitude} instance.
}
\description{
With this method you can create a longitude axis to use with new \link{CFArray}
instances.
}
