% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeLongitudeAxis}
\alias{makeLongitudeAxis}
\title{Create a longitude axis}
\usage{
makeLongitudeAxis(id, name, group, length, values, bounds = NULL, units = "")
}
\arguments{
\item{id}{Id of the axis.}

\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{length}{Length of the dimension of the axis.}

\item{values}{The dimension values.}

\item{bounds}{The bounds of the dimension values, or \code{NULL} if not available.}

\item{units}{A character string with the axis units.}
}
\value{
A \code{CFAxisLongitude} instance.
}
\description{
With this method you can create a longitude axis to use with new \link{CFArray}
instances.
}
