% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeMemoryGroup}
\alias{makeMemoryGroup}
\title{Create a group in memory to hold CF objects}
\usage{
makeMemoryGroup(id, name, fullname, title, history)
}
\arguments{
\item{id}{The id of the group.}

\item{name}{The name of the group.}

\item{fullname}{The full path and name of the group.}

\item{title}{The title attribute of the group.}

\item{history}{The history attribute of the group.}
}
\value{
A \code{MemoryGroup} instance.
}
\description{
With this function a group is created in memory, i.e. not associated with a
netCDF resource on file. This can be used to prepare new CF objects before
writing them to file. Extracting data from a \code{CFVariable} into a \link{CFData}
instance will also create a memory group.
}
