% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCFObjects.R
\name{makeTimeAxis}
\alias{makeTimeAxis}
\title{Create a time axis}
\usage{
makeTimeAxis(id, name, group, values)
}
\arguments{
\item{id}{Id of the axis.}

\item{name}{Name of the axis.}

\item{group}{Group to place the axis in.}

\item{values}{A \code{CFTime} instance with time values and bounds set.}
}
\value{
A \code{CFAxisTime} instance.
}
\description{
With this method you can create a time axis to use with new \link{CFArray}
instances.
}
