% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfDimensionNumeric.R
\docType{class}
\name{ncdfDimensionNumeric-class}
\alias{ncdfDimensionNumeric-class}
\title{Numeric dimension class}
\description{
This class describes numeric dimensions in an NetCDF resource.
}
\section{Slots}{

\describe{
\item{\code{values}}{The values of the positions along the dimension.}

\item{\code{bounds}}{The bounds of the dimension values, if any.}
}}

