% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfVariable.R
\docType{class}
\name{ncdfVariable-class}
\alias{ncdfVariable-class}
\title{The ncdfVariable class}
\description{
The ncdfVariable class
}
\section{Slots}{

\describe{
\item{\code{dims}}{\code{ncdfDimension}. Vector of dimensions of this variable.}
}}

